add: route mock api Restaurants#45
